updated install_aws_cli.sh

+++ setup/install_aws_cli.sh
@@ -81,6 +81,10 @@ else
         echo
         brew install awless
     else
+        if ! curl -s https://updates.awless.io >/dev/null; then
+            echo "AWLess SSL certificate still expired, must install manually until fixed"
+            exit 0
+        fi
         curl https://raw.githubusercontent.com/wallix/awless/master/getawless.sh | bash
         mv -iv awless ~/bin/
     fi
